global state e 'carregando' pra esperar as requisições
