+++ pokedex/src/App.js
@@ -1,8 +1,16 @@
 import React from "react";
 import Router from "./routes/Router";
+import { GlobalState } from "./global/GlobalState";
+
+export const GlobalStateContext = React.createContext()
 
 const App = () => {
-  return <Router />;
+
+  return (
+    <GlobalState>
+      <Router />
+    </GlobalState>
+  );
 };
 
 export default App;
