+++ pokedex/src/components/PokeCard/PokeCard.js
@@ -1,81 +1,56 @@
-import axios from "axios";
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect } from "react";
 import { useHistory } from "react-router-dom";
-import { BASE_URL } from "../../constants/urls";
+import { GlobalStateContext } from "../../global/GlobalStateContext";
 import { goToDetails } from "../../routes/coordinator";
 import { CardsContainer, PokemonCard, CardButtonsContainer, CardButtonLeft, CardButtonRight } from "./styled";
 
 const PokeCard = (props) => {
     const history = useHistory();
-    const [page, setPage] = useState("")
-
-    const [pokeList, setPokeList] = useState([])
-    const [pokeDetail, setPokeDetail] = useState([])
-
-    const getPokemon = () => {
-        axios.get(`${BASE_URL}/?limit=40`)
-            .then((res) => {
-                setPokeList(res.data.results)
-            })
-            .catch((err) => {
-                alert(err.response)
-            })
-    }
-
-    const getDetail = (pokeList) => {
-        const detailList = []
-
-        pokeList.forEach((poke) => {
-            axios.get(`${BASE_URL}/${poke.name}`).then((res) => {
-                detailList.push(res.data)
-                if (detailList.length === 40) {
-                    const pokeOrder = detailList.sort((a, b) => {
-                        return a.id - b.id
-                    })
-                    setPokeDetail(pokeOrder)
-                }
-            })
-        });
-    }
+    const { pokedex, pokeList, pokeDetail, page, setPage, addPokemonToPokedex, removePokemonToPokedex, getPokemonList, getPokemonListWithImages } = useContext(GlobalStateContext)
 
     useEffect(() => {
-        getPokemon()
+        getPokemonList()
         setPage(props.currentPage)
     }, [setPage, props.currentPage])
 
     useEffect(() => {
-        getDetail(pokeList)
+        getPokemonListWithImages(pokeList)
     }, [pokeList])
 
     const renderCardPage = () => {
 
         if (page === "home") {
-            return pokeDetail && pokeDetail.map((poke) => {
-                return <PokemonCard key={poke.name}>
-                    <p>#{poke.id} <strong>{poke.name}</strong></p>
-                    <img alt={poke.name} src={poke.sprites.front_default} />
-                    <CardButtonsContainer>
-                        <CardButtonLeft onClick={() => props.addPokemonToPokedex(poke)}>Adicionar ao Pokédex</CardButtonLeft>
-                        <CardButtonRight onClick={() => goToDetails(history, poke.name)}>Ver detalhes</CardButtonRight>
-                    </CardButtonsContainer>
-                </PokemonCard>
-            })
+            return pokeDetail[0] ? (
+                pokeDetail.map((poke) => {
+                    return <PokemonCard key={poke.name}>
+                        <p>#{poke.id} <strong>{poke.name}</strong></p>
+                        <img alt={poke.name} src={poke.sprites.front_default} />
+                        <CardButtonsContainer>
+                            <CardButtonLeft onClick={() => addPokemonToPokedex(poke)}>Adicionar ao Pokédex</CardButtonLeft>
+                            <CardButtonRight onClick={() => goToDetails(history, poke.name)}>Ver detalhes</CardButtonRight>
+                        </CardButtonsContainer>
+                    </PokemonCard>
+                })
+            ) : (
+                <p>Carregando...</p>
+            )
 
         } else if (page === "pokedex") {
-            return props.pokedex && props.pokedex.map((poke) => {
+            return pokedex[0] ? (pokedex.map((poke) => {
                 return <PokemonCard key={poke.name}>
                     <p>#{poke.id} <strong>{poke.name}</strong></p>
                     <img alt={poke.name} src={poke.sprites.front_default} />
                     <CardButtonsContainer>
-                        <CardButtonLeft onClick={() => props.removePokemonToPokedex(poke)}>Remover do Pokédex</CardButtonLeft>
+                        <CardButtonLeft onClick={() => removePokemonToPokedex(poke)}>Remover do Pokédex</CardButtonLeft>
                         <CardButtonRight onClick={() => goToDetails(history, poke.name)}>Ver detalhes</CardButtonRight>
                     </CardButtonsContainer>
                 </PokemonCard>
             })
+            ) : (
+                <p>Seu Pokédex ainda está vazio :(</p>
+            )
         }
     }
-
-
 
     return (
         <CardsContainer>
