+++ pokedex/src/components/PokeCard/styled.js
@@ -3,11 +3,9 @@ import styled from "styled-components";
 
 export const CardsContainer = styled.div`
     display:flex;
-    /* margin: 16px; */
     flex-wrap:wrap;
     width:80%;
     justify-content:center;
-    /* border-radius:8px; */
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
     box-sizing:border-box;
 `
