+++ pokedex/src/global/GlobalStateContext.js
@@ -1,0 +1,3 @@
+import React from "react"
+
+export const GlobalStateContext = React.createContext()
