+++ pokedex/src/pages/PokedexPage/PokedexPage.js
@@ -1,29 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import Logo from "../../img/Logo.png";
-import { ImgContainer, Img, PokemonListContainer } from "./styled";
+import { ImgContainer, Img, PokemonListContainer, StyledButton } from "./styled";
 import '../../App.css';
 import { useHistory } from "react-router-dom";
 import { goToHome, goToPokedex } from "../../routes/coordinator";
-import { withStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
 import PokeCard from "../../components/PokeCard/PokeCard";
-
-const StyledButton = withStyles({
-  root: {
-    background: 'linear-gradient(45deg, #282c34 30%, #444b59 90%)',
-    borderRadius: 3,
-    border: 0,
-    color: 'white',
-    height: 36.5,
-    padding: '0 15px',
-    marginLeft: '70px',
-    boxShadow: '0 3px 5px 2px rgba(255, 255, 255, .3)',
-    transition: '0.3s ease-in-out',
-  },
-  label: {
-    textTransform: 'capitalize',
-  },
-})(Button);
 
 const PokedexPage = (props) => {
   const history = useHistory()
@@ -47,9 +28,6 @@ const PokedexPage = (props) => {
 
         <PokeCard
           currentPage={currentPage}
-          pokedex={props.pokedex}
-          setPokedex={props.setPokedex}
-          removePokemonToPokedex={props.removePokemonToPokedex}
         />
 
       </PokemonListContainer>
